requirements.txt
ahk
ahk[binary]
python-dotenv
discord.py